// src/lib/openapi-signalement/models/ToponymeChangesRequestedDTO.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { PositionDTO } from './PositionDTO';

export type ToponymeChangesRequestedDTO = {
    nom: string;
    parcelles: Array<string>;
    positions: Array<PositionDTO>;
    comment?: string | null;
};